Split out Static API handler tests
